Fix 'Cannot access before initialization' crash by breaking circular import
plannerObjectiveScoring imported metric functions from autoPlannerOptimizationEngine,
which imported evaluatePlannerObjectives from plannerObjectiveEngine, which imported
from plannerObjectiveScoring — a three-way circular dependency causing a TDZ crash
at runtime (minified as 'vs').

Fix: extract the pure metric functions (calculateMealFatigueScore,
calculateWeeklyPlannerStress, calculateIngredientOverlapScore,
calculateGroceryFragmentation, calculatePantryReuseEfficiency, etc.) into a new
autoPlannerMetrics.ts that has no dependency on plannerObjectiveEngine. Both
autoPlannerOptimizationEngine and plannerObjectiveScoring now import from
autoPlannerMetrics — no cycle. Re-enable ENABLE_JOURNEY_TIMELINE = true.
